+++ crypto/stark/src/lib.rs
@@ -88,7 +88,10 @@ pub use verifier::{verify, Error as VerifierError};
 // Exports for prover
 #[cfg(feature = "prover")]
 pub use component::{
-    compose_folded, compose_horizontal, compose_vertical, fold, fold_many, permute_columns, shift,
+    compose::{
+        compose_folded, compose_horizontal, compose_vertical, fold, fold_many, permute_columns,
+        shift,
+    },
     Component,
 };
 #[cfg(feature = "prover")]
